requirements.txt
streamlit
requests>=2.28
beautifulsoup4>=4.12
python-dotenv>=0.21.0
google-genai>=0.0.0
pandas>=2.0
plotly>=5.0
